+++ pkg/evaluator/builtins/std.go
@@ -19,7 +19,7 @@ var bblen = &object.Builtin{
 		}
 		if val, ok := args[0].(*object.String); ok {
 			return &object.Integer{
-				Value: int32(len(val.Value)),
+				Value: int64(len(val.Value)),
 			}
 		}
 		return newError("expected %s, got %s", object.STRING, args[0].Type())
